+++ docs/bundle-analyzer.config.js
@@ -1,0 +1,8 @@
+const BundleAnalyzerPlugin = require('webpack-bundle-analyzer')
+  .BundleAnalyzerPlugin
+
+console.log('bundle analyzer')
+
+module.exports = {
+  plugins: [new BundleAnalyzerPlugin()],
+}