+++ packages/components/src/MDXDeck.js
@@ -13,6 +13,7 @@ import Print from './Print'
 import GoogleFonts from './GoogleFonts'
 import Catch from './Catch'
 import Keyboard from './Keyboard'
+import Storage from './Storage'
 
 const NORMAL = 'normal'
 const PRESENTER = 'presenter'
@@ -26,9 +27,6 @@ const modes = {
   GRID,
   PRINT,
 }
-
-const STORAGE_INDEX = 'mdx-slide'
-const STORAGE_STEP = 'mdx-step'
 
 const BaseWrapper = props => <>{props.children}</>
 
@@ -107,22 +105,6 @@ export class MDXDeck extends React.Component {
     this.setState({ slides })
   }
 
-  handleStorageChange = e => {
-    const { key } = e
-    switch (key) {
-      case STORAGE_INDEX:
-        const index = parseInt(e.newValue, 10)
-        this.goto(index)
-        break
-      case STORAGE_STEP:
-        const step = parseInt(e.newValue, 10)
-        this.setState({ step })
-        break
-      default:
-        break
-    }
-  }
-
   getMode = () => {
     const query = querystring.parse(
       globalHistory.location.search.replace(/^\?/, '')
@@ -131,20 +113,12 @@ export class MDXDeck extends React.Component {
   }
 
   componentDidMount() {
-    window.addEventListener('storage', this.handleStorageChange)
     this.getMode()
   }
 
-  componentWillUnmount() {
-    window.removeEventListener('storage', this.handleStorageChange)
-  }
-
   componentDidUpdate() {
-    const index = this.getIndex()
-    const { step, mode } = this.state
+    const { mode } = this.state
     const { pathname, search } = globalHistory.location
-    localStorage.setItem(STORAGE_INDEX, index)
-    localStorage.setItem(STORAGE_STEP, step)
 
     if (mode !== NORMAL && mode !== PRINT) {
       const query = '?' + querystring.stringify({ mode })
@@ -209,6 +183,7 @@ export class MDXDeck extends React.Component {
         {style}
         <Catch>
           <Keyboard {...this.props} {...context} />
+          <Storage {...this.state} goto={this.goto} index={index} />
           <GoogleFonts />
           <Wrapper {...this.props} {...this.state} modes={modes} index={index}>
             <Swipeable onSwipedRight={this.previous} onSwipedLeft={this.next}>