+++ packages/components/src/Storage.js
@@ -1,0 +1,44 @@
+import { useEffect } from 'react'
+
+const STORAGE_INDEX = 'mdx-slide'
+const STORAGE_STEP = 'mdx-step'
+
+export const useLocalStorage = (handler, args = []) => {
+  useEffect(() => {
+    window.addEventListener('storage', handler)
+    return () => {
+      window.removeEventListener('storage', handler)
+    }
+  }, [...args])
+}
+
+export const useSetStorage = (key, value) => {
+  useEffect(() => {
+    localStorage.setItem(key, value)
+  }, [key, value])
+}
+
+const handleStorageChange = ({ goto, update }) => e => {
+  const { key } = e
+  switch (key) {
+    case STORAGE_INDEX:
+      const index = parseInt(e.newValue, 10)
+      goto(index)
+      break
+    case STORAGE_STEP:
+      const step = parseInt(e.newValue, 10)
+      update({ step })
+      break
+    default:
+      break
+  }
+}
+
+export default ({ goto, update, index, step }) => {
+  const handler = handleStorageChange({ goto, update })
+  useLocalStorage(handler)
+  useSetStorage(STORAGE_INDEX, index)
+  useSetStorage(STORAGE_STEP, step)
+
+  return false
+}